moved functinos with pointers to its own source file
